Fix order-of-occurence problems: depth-first + emit latest Dyn value
Breadth-first solved some ordering problems while introducing others.
In the end, neither breadth-first nor depth-first solved the real
problem, which was that the first subscriber to a Dynamic could in turn
end up modifying the dynamic, but subsequent subscribers to that same
Dynamic would end up getting the old value, not the new value.

Having Dynamics emit a thunk, so that the value to be emitted it always
calculated based on the time of running *that particular* subscriber
ensures that subscribers always get the latest value, preventing
subsequent subscribers from overriding effects from prior subscribers.

This hopefully solves the order-of-occurrence problems once and for all,
but without the resources for a proper deep-dive into the semantics of
the FRP system there is still a chance of other problems cropping up
elsewhere.