Ensure rendering code only runs once per frame
Previously rendering code was run whenever the concatDyn emitted, which
would emit any time the ship or the asteroids or the explosions were
updated - which would be many times per frame. To ensure that rendering
only occurs once per frame, we instead need to *sample* the concatDyn
when a *frame* occurs, locking rendering to once per frame.